spectrogram plugin: move to separate directory (#1996)
* separate spectrogram plugin into multiple files

* add changelog entry
